(trunk) Fixed exception when subtracting two 32-bit unsigned integers and getting what should be a negative result, but instead getting large positive number near max 32-bit. It's because storing in a 64-bit signed integer (it works if storing in a 32-bit signed integer though) - another reason to use signed integers in general instead of unsigned. Also using a static_cast<int>() in case unsigned quantities are changed to non-32-bit types in future for some reason.
git-svn-id: http://svn.gplates.org/projects/gplates/trunk@19162 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c

+++ src/maths/PolygonOnSphere.cc
@@ -1010,7 +1010,8 @@ GPlatesMaths::PolygonOnSphere::ConstIterator::distance_to(
 			GPLATES_ASSERTION_SOURCE,
 			"Attempted to compare an uninitialised iterator.");
 
-	const int ring_id_difference = other.d_current_ring_id - d_current_ring_id;
+	const int ring_id_difference =
+			static_cast<int>(other.d_current_ring_id) - static_cast<int>(d_current_ring_id);
 
 	if (ring_id_difference == 0)
 	{
